requirements.txt
# Core PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
pdf2image==1.16.3

# Data Processing
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2

# NLP and Text Matching
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Web Framework
streamlit==1.28.2
plotly==5.18.0
streamlit-aggrid==0.3.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
colorlog==6.7.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0